fix(core): harden retrieval comparison telemetry
